Feature/설문항목생성

+++ src/main/java/org/jenga/dantong/post/service/PostService.java
@@ -1,9 +1,6 @@
 package org.jenga.dantong.post.service;
 
 import jakarta.transaction.Transactional;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.jenga.dantong.global.s3.model.dto.FileUploadRequest;
@@ -28,6 +25,10 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
+import java.util.ArrayList;
+import java.util.List;
+import java.util.stream.Collectors;
+
 @Slf4j
 @Service
 @RequiredArgsConstructor
@@ -38,17 +39,17 @@ public class PostService {
     private final FileUploadService fileUploadService;
 
     @Transactional
-    public int savePost(PostCreateRequest request, Long userId) {
+    public Long savePost(PostCreateRequest request, Long userId) {
         User user = userRepository.findById(userId).orElseThrow(UserNotFoundException::new);
         Post post = Post.builder()
-            .user(user)
-            .title(request.getTitle())
-            .description(request.getDescription())
-            .content(request.getContent())
-            .category(request.getCategory())
-            .startDate(request.getStartTime())
-            .endDate(request.getEndTime())
-            .build();
+                .user(user)
+                .title(request.getTitle())
+                .description(request.getDescription())
+                .content(request.getContent())
+                .category(request.getCategory())
+                .startDate(request.getStartTime())
+                .endDate(request.getEndTime())
+                .build();
         log.info(request.getTitle());
         if (request.getImageFiles() != null) {
             saveFiles(request.getImageFiles(), post);
@@ -58,19 +59,19 @@ public class PostService {
     }
 
     @Transactional
-    public PostResponse findPost(int postId) {
+    public PostResponse findPost(Long postId) {
         Post post = postRepository.findById(postId).orElseThrow(PostNofFoundException::new);
         String progress = Util.getProgress(post);
         List<PostFileResponse> files = post.getFiles().stream()
-            .map(file -> {
-                String url = fileUploadService.getFileUrl(file.getFileId());
-                return new PostFileResponse(file, url);
-            }).collect(Collectors.toList());
+                .map(file -> {
+                    String url = fileUploadService.getFileUrl(file.getFileId());
+                    return new PostFileResponse(file, url);
+                }).collect(Collectors.toList());
         return new PostResponse(post, progress, files);
     }
 
     @Transactional
-    public int deletePost(int postId) {
+    public Long deletePost(Long postId) {
         Post post = postRepository.findById(postId).orElseThrow(PostNofFoundException::new);
         post.setShown(false);
         postRepository.save(post);
@@ -93,19 +94,19 @@ public class PostService {
     private Page<PostResponse> getPostResponses(Page<Post> posts) {
         return posts.map(currPost -> {
             List<PostFileResponse> files = currPost.getFiles().stream()
-                .map(file -> {
-                    String url = fileUploadService.getFileUrl(file.getFileId());
-                    return new PostFileResponse(file, url);
-                }).collect(Collectors.toList());
+                    .map(file -> {
+                        String url = fileUploadService.getFileUrl(file.getFileId());
+                        return new PostFileResponse(file, url);
+                    }).collect(Collectors.toList());
             String progress = Util.getProgress(currPost);
             return new PostResponse(currPost, progress, files);
         });
     }
 
     @Transactional
-    public int updatePost(PostUpdateRequest request, Long userId) {
+    public Long updatePost(PostUpdateRequest request, Long userId) {
         Post post = postRepository.findById(request.getPostId())
-            .orElseThrow(PostNofFoundException::new);
+                .orElseThrow(PostNofFoundException::new);
         if (!userId.equals(post.getUser().getId())) {
             throw new PermissionDeniedException();
         }
@@ -121,14 +122,14 @@ public class PostService {
 
     private void saveFiles(List<MultipartFile> files, Post post) {
         List<RequestFile> requestFiles = fileUploadService.uploadFiles(
-            FileUploadRequest.ofList(files));
+                FileUploadRequest.ofList(files));
         List<PostFile> postFiles = new ArrayList<>();
 
         for (RequestFile file : requestFiles) {
             PostFile.PostFileBuilder builder = PostFile.builder()
-                .fileName(file.getOriginalName())
-                .mediaType(file.getMediaType().toString())
-                .fileId(file.getFileId());
+                    .fileName(file.getOriginalName())
+                    .mediaType(file.getMediaType().toString())
+                    .fileId(file.getFileId());
 
             postFiles.add(builder.build());
         }
